pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "block-detect"
version = "0.1.0"
description = "Batch image block detection for BMotion captures"
requires-python = ">=3.11"
dependencies = [
  "dropbox>=11,<13",
  "Pillow>=12,<13",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
